+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added missing `Nanook::WalletAccount#block_count` delegate.
+- Added `Nanook#network_telemetry`.
 - Added `Nanook::WalletAccount#work`.
 - Added `Nanook::WalletAccount#set_work`.
 - Added `Nanook::Account#delegators_count`.